requirements.txt
argparse
Jinja2
PyYAML